connectionResetError (NEEDS FIX)
only trying to download 5 folders of images rn ~500 jps from each but runs into ConnectionResetError/ProtocolError.ChunckedEncodingError every 1.5-2 files.
Added code to make sure it doesnt try to download an image that already exists but fixing the connection error would be more ideal

+++ pillImageDownload.py
@@ -3,6 +3,7 @@ import requests
 import re
 import os
 import urllib.request
+from os.path import exists
 
 def folder_create(images, url):
     folder_name = "images"
@@ -14,14 +15,15 @@ def download_images(images,url, folder_name): #, folder_name):
     if len(images) != 0:
         for i, image in enumerate(images):
             image_link = url + "/" + image
-            r = requests.get(image_link).content
-            with open(f"{folder_name}/{image}", "wb+") as f:
-                f.write(r)
-                count+=1
-            
-            #keeping track of image downloads
-            if count%100 == 0:
-                print(f"{count} images downloaded from {url}")
+            if not exists(f"{folder_name}/{image}"):
+                r = requests.get(image_link).content
+                with open(f"{folder_name}/{image}", "wb+") as f:
+                    f.write(r)
+                    count+=1
+                
+                #keeping track of image downloads
+                if count%100 == 0:
+                    print(f"{count} images downloaded from {url}")
 
         if count == len(images):
             print("All the images have been downloaded!")
@@ -56,6 +58,10 @@ for link in soup.find_all("a"):
     if link.get("href").startswith("PillProjectDisc"):
         links.append(link.get("href"))
 
+curr = [l.strip('/index.html') for l in links[:5]]
+for c in curr:
+    print(c)
+
 for l in range(len(links)):
     if links[l].endswith('/index.html'):
         links[l] = links[l][:-11]
@@ -63,6 +69,8 @@ for l in range(len(links)):
 
 # for l in links:
 #     print(l)
-folder_name = "images"
-os.mkdir(folder_name)
-main(links[:5])
+#folder_name = "images"
+#os.mkdir(folder_name)
+#print(links[:5])
+#print(links[2:5])
+main(links[3:5])
